Done base functionality

+++ src/Controller/TransactionController.php
@@ -30,7 +30,7 @@ class TransactionController extends AbstractController
      */
     public function new(Request $request, UserLoader $userLoader, TradeMaster $tradeMaster) : Response
     {
-        $this->denyAccessUnlessGranted('CREATE_TRANSACTION');
+        $this->denyAccessUnlessGranted('CREATE_TRANSACTIONS');
 
         $currentUser = $userLoader->getUser();
         $transaction = (new Transaction())
@@ -61,6 +61,8 @@ class TransactionController extends AbstractController
      */
     public function show(Transaction $transaction) : Response
     {
+        $this->denyAccessUnlessGranted('SHOW', $transaction);
+
         return $this->render('transaction/show.html.twig', ['transaction' => $transaction]);
     }
 
@@ -69,6 +71,8 @@ class TransactionController extends AbstractController
      */
     public function edit(Request $request, Transaction $transaction) : Response
     {
+        $this->denyAccessUnlessGranted('EDIT', $transaction);
+
         $form = $this->createForm(TransactionType::class, $transaction);
         $form->handleRequest($request);
 
@@ -89,6 +93,8 @@ class TransactionController extends AbstractController
      */
     public function delete(Request $request, Transaction $transaction) : Response
     {
+        $this->denyAccessUnlessGranted('DELETE', $transaction);
+
         if ($this->isCsrfTokenValid('delete' . $transaction->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($transaction);
